Fix Opera Wrocław scraper to correctly identify performances
Previously the scraper would traverse too far up the parent tree when
looking for dates, causing it to incorrectly associate performances
with dates from other shows on the page. Now it finds div.rep-single
containers which contain both the date (YYYYMMDD) AND the opera title,
ensuring each performance is correctly matched with its own date.

This fixes the bug where "Halka" was incorrectly shown on Dec 31, 2025
(which is actually "Zemsta nietoperza").